Delected the accidentally created duplicate
When merging the branch created by Sebastian 4 months ago, I missed out that it was using the previous name of the folder. So it created a duplicate rather a merge. I deleted it and saved the content to some space on my computer (in case it will be somehow needed)